+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/TableTestExtension.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2026 DiffPlug
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.diffplug.gradle.spotless;
+
+import javax.inject.Inject;
+
+import com.diffplug.spotless.FormatterStep;
+import com.diffplug.spotless.java.TableTestFormatterStep;
+
+public class TableTestExtension extends FormatExtension {
+	static final String NAME = "tableTest";
+
+	@Inject
+	public TableTestExtension(SpotlessExtension spotless) {
+		super(spotless);
+	}
+
+	@Override
+	protected void setupTask(SpotlessTask task) {
+		if (target == null) {
+			throw noDefaultTargetException();
+		}
+		super.setupTask(task);
+	}
+
+	public TableTestFormatterConfig tableTestFormatter() {
+		return tableTestFormatter(TableTestFormatterStep.defaultVersion());
+	}
+
+	public TableTestFormatterConfig tableTestFormatter(String version) {
+		return new TableTestFormatterConfig(version);
+	}
+
+	public class TableTestFormatterConfig {
+		private final String version;
+
+		TableTestFormatterConfig(String version) {
+			this.version = version;
+			addStep(createStep());
+		}
+
+		private FormatterStep createStep() {
+			return TableTestFormatterStep.create(version, provisioner());
+		}
+	}
+}